闹钟刷新成功，加入报时

+++ AngoTime/AngoTime/AngoTimeDlg.cpp
@@ -8,12 +8,23 @@
 #include "afxdialogex.h"
 #include "MsgBoxEx.h"
 
+//playsound
+#include <mmsystem.h>
+#pragma comment( lib, "Winmm.lib" )
+
 #define PI 3.1415926535897932384626433832795028841971693993751058209
 
 #ifdef _DEBUG
 #define new DEBUG_NEW
 #endif
 
+
+
+CWinThread*			g_pthAlarm		=	NULL;
+CWinThread*			g_pthSayTime	=	NULL;
+CWinThread*			g_pthWork		=	NULL;
+BOOL				g_bWork			=	TRUE;
+DWORD				g_dwTasktype	=	0;
 
 // CAngoTimeDlg 对话框
 
@@ -24,10 +35,18 @@ CAngoTimeDlg::CAngoTimeDlg(CWnd* pParent /*=NULL*/)
 	, m_uClock_Timer(0)
 	, m_Point_Start(0)
 	, m_Point_End(0)
-	, m_s(0)
-	, m_m(0)
-	, m_h(0)
 {
+	m_point_lastHour	=	0;
+	m_point_lastMin		=	0;
+	m_point_lastSecL	=	0;
+	m_point_lastSecS	=	0;
+	m_bFirstClock		=	TRUE;
+
+	m_cLastHour		=	0;
+	m_cLastMin		=	0;
+	m_cLastSecL		=	0;
+	m_cLastSecS		=	0;
+
 	m_hIcon = AfxGetApp()->LoadIcon(IDR_MAINFRAME);
 	m_popMenu.LoadMenu(IDR_MENU_RBTN);	
 }
@@ -53,10 +72,29 @@ BEGIN_MESSAGE_MAP(CAngoTimeDlg, CDialogEx)
 	ON_COMMAND(ID_VIEW_SHOW, &CAngoTimeDlg::OnViewShow)
 	ON_COMMAND(ID_VIEW_HIDE, &CAngoTimeDlg::OnViewHide)
 	ON_COMMAND(ID_MENU_ANGO, &CAngoTimeDlg::OnMenuAngo)
+	ON_WM_CLOSE()
+	ON_COMMAND(ID_SAYTIME_NOW, &CAngoTimeDlg::OnSaytimeNow)
 END_MESSAGE_MAP()
 
 
 // CAngoTimeDlg 消息处理程序
+void CAngoTimeDlg::OnClose()
+{
+	// TODO:  在此添加消息处理程序代码和/或调用默认值
+
+	g_bWork = FALSE;
+	g_pthAlarm->ResumeThread();
+	g_pthSayTime->ResumeThread();
+	g_pthWork->ResumeThread();
+
+	WaitForSingleObject(g_pthAlarm->m_hThread, INFINITE);
+	WaitForSingleObject(g_pthSayTime->m_hThread, INFINITE);
+	WaitForSingleObject(g_pthWork->m_hThread, INFINITE);
+
+
+	CDialogEx::OnClose();
+	CDialogEx::OnCancel();
+}
 
 BOOL CAngoTimeDlg::OnInitDialog()
 {
@@ -97,6 +135,7 @@ BOOL CAngoTimeDlg::OnInitDialog()
 	OnViewShow();
 	OnViewUp();
 	
+	InitThread();
 
 	return TRUE;  // 除非将焦点设置到控件，否则返回 TRUE
 }
@@ -182,20 +221,11 @@ void CAngoTimeDlg::OnRButtonDown(UINT nFlags, CPoint point)
 
 void CAngoTimeDlg::OnOK()
 {
-	CString strMsg = L"回车键";
-	AngoMessageBox(strMsg);
-
 }
 
 void CAngoTimeDlg::OnCancel()
 {
-	CString strMsg;
-	strMsg = "是否退出程序？";
-	if (AngoMessageBox(strMsg, L"退出", MB_OKCANCEL) == IDCANCEL)
-	{
-		return;
-	}
-	CDialog::OnCancel();
+	OnClose();
 }
 
 //-------------------------------------------------------------------------------------------------------------------------
@@ -222,13 +252,6 @@ LRESULT CAngoTimeDlg::OnNotifyIcon(WPARAM wparam, LPARAM lparam)
 
 
 //-------------------------------------------------------------------------------------------------------------------------
-
-void CAngoTimeDlg::OnMenuExit()
-{
-	OnCancel();
-}
-
-//-------------------------------------------------------------------------------------------------------------------------
 void CAngoTimeDlg::InitClock()
 {
 	//设置控制时针走动的触发器为每秒一次
@@ -238,7 +261,7 @@ void CAngoTimeDlg::InitClock()
 	CRgn  rgn;
 	CRect  rc;
 	GetClientRect(&rc);
-	rgn.CreateEllipticRgn(rc.left, rc.top, rc.right, rc.bottom);
+	rgn.CreateEllipticRgn(rc.left, rc.top, rc.right-1, rc.bottom-1);
 	SetWindowRgn(rgn, TRUE);
 	rgn.DeleteObject();
 	
@@ -256,79 +279,53 @@ HBRUSH CAngoTimeDlg::OnCtlColor(CDC* pDC, CWnd* pWnd, UINT nCtlColor)
 	return m_cBrush;
 }
 
-void  CAngoTimeDlg::GetRgb(int &nPos, tagRGB &tRgb)
-{
-	int C = 0;
-
-	if (nPos < 5)
-		C = 250;
-	else if (nPos < 10)
-		C = 247;
-	else if (nPos < 15)
-		C = 244;
-	else if (nPos < 20)
-		C = 240;
-	else if (nPos < 30)
-		C = 235;
-	else if (nPos < 35)
-		C = 235;
-	else if (nPos < 40)
-		C = 240;
-	else if (nPos < 45)
-		C = 243;
-	else if (nPos < 50)
-		C = 245;
-	else
-		C = 250;
-
-	tRgb.x = C;
-	tRgb.y = C;
-	tRgb.z = C;
-
-}
-
-void  CAngoTimeDlg::GetRgb(CPoint &cPoint, tagRGB &tRgb)
-{
-
-}
 
 
 void  CAngoTimeDlg::ClockTime()
 {
 	//获得当前系统时间。
 	CTime time = CTime::GetCurrentTime();
-	tagRGB tRgb = { 0 };
 	CPen *pPenOld = NULL;
 	CPen PenNew;
 	CBrush *pBrushOld = NULL;
 	CBrush BrushNew;
 	CClientDC dc(this);
+
 	int   S = time.GetSecond();
 	float M = float(time.GetMinute() + S / 60.0);
 	float H = float(time.GetHour() + M / 60.0);
 	if (H > 12)
 		H = H - 12;
 	H = H * 5;
+
 	m_Point_Start.x = 65;
 	m_Point_Start.y = 64;
+	COLORREF cRgb	= 0;
 
-
-	//以下为画时针分针和秒针
+	
 	//方法为画每根针前用背景色擦去上一次画的针（由于背景色渐变，所以加入了计算）
 	//从图片获取背景色rgb
+	//COLORREF cRgb = dc.GetPixel(m_Point_End);
+	//BYTE byRed   = GetRValue(cRgb);
+	//BYTE byGreen = GetGValue(cRgb);
+	//BYTE byBlue	 = GetBValue(cRgb);
+
 	//////////////////////////////////////////////	
-	GetRgb(m_s, tRgb);
+	if ( !m_bFirstClock )
+	{
+		//用背景色擦去上一次画的针
+		cRgb = m_cLastHour;
+		m_Point_End = m_point_lastHour;
 
-	PenNew.CreatePen(PS_SOLID, 4, RGB(tRgb.x, tRgb.y, tRgb.z));
-	BrushNew.CreateSolidBrush(RGB(tRgb.x, tRgb.y, tRgb.z));
-	pBrushOld = dc.SelectObject(&BrushNew);
-	pPenOld = dc.SelectObject(&PenNew);
-	m_Point_End.x = 65 + LONG(22 * sin(m_h*PI / 30));
-	m_Point_End.y = 64 - LONG(22 * cos(m_h*PI / 30));
-	dc.MoveTo(m_Point_Start);
-	dc.LineTo(m_Point_End);
+		PenNew.CreatePen(PS_SOLID, 4, cRgb);
+		BrushNew.CreateSolidBrush(cRgb);
+		pBrushOld = dc.SelectObject(&BrushNew);
+		pPenOld = dc.SelectObject(&PenNew);
+		dc.MoveTo(m_Point_Start);
+		dc.LineTo(m_Point_End);
+	}
 
-
+	//画时针
 	PenNew.DeleteObject();
 	PenNew.CreatePen(PS_SOLID, 4, RGB(0, 0, 0));
 	pPenOld = dc.SelectObject(&PenNew);
@@ -337,24 +334,34 @@ void  CAngoTimeDlg::ClockTime()
 	pBrushOld = dc.SelectObject(&BrushNew);
 	m_Point_End.x = 65 + LONG(22 * sin(H*PI / 30));
 	m_Point_End.y = 64 - LONG(22 * cos(H*PI / 30));
+
+	//保存原有的rgb
+	m_cLastHour = dc.GetPixel(m_Point_End);
+	m_point_lastHour = m_Point_End;
+
 	dc.MoveTo(m_Point_Start);
 	dc.LineTo(m_Point_End);
+	
+
+
 	///////////////////////////////////////////////
-	GetRgb(m_s, tRgb);
+	if ( !m_bFirstClock )
+	{
+		//擦除上次的指针
+		cRgb = m_cLastMin;
+		m_Point_End = m_point_lastMin;
 
-	BrushNew.DeleteObject();
-	BrushNew.CreateSolidBrush(RGB(tRgb.x, tRgb.y, tRgb.z));
-	pBrushOld = dc.SelectObject(&BrushNew);
-	PenNew.DeleteObject();
-	PenNew.CreatePen(PS_SOLID, 3, RGB(tRgb.x, tRgb.y, tRgb.z));
-	pPenOld = dc.SelectObject(&PenNew);
-	m_Point_End.x = 65 + LONG(30 * sin(m_m*PI / 30));
-	m_Point_End.y = 64 - LONG(30 * cos(m_m*PI / 30));
-	dc.MoveTo(m_Point_Start);
-	dc.LineTo(m_Point_End);
+		BrushNew.DeleteObject();
+		BrushNew.CreateSolidBrush(cRgb);
+		pBrushOld = dc.SelectObject(&BrushNew);
+		PenNew.DeleteObject();
+		PenNew.CreatePen(PS_SOLID, 3, cRgb);
+		pPenOld = dc.SelectObject(&PenNew);
+		dc.MoveTo(m_Point_Start);
+		dc.LineTo(m_Point_End);
+	}
 
-
-
+	//画分针
 	BrushNew.DeleteObject();
 	BrushNew.CreateSolidBrush(RGB(0, 0, 0));
 	pBrushOld = dc.SelectObject(&BrushNew);
@@ -363,49 +370,67 @@ void  CAngoTimeDlg::ClockTime()
 	pPenOld = dc.SelectObject(&PenNew);
 	m_Point_End.x = 65 + LONG(30 * sin(M*PI / 30));
 	m_Point_End.y = 64 - LONG(30 * cos(M*PI / 30));
+
+	m_cLastMin = dc.GetPixel(m_Point_End);
+	m_point_lastMin = m_Point_End;
+
 	dc.MoveTo(m_Point_Start);
 	dc.LineTo(m_Point_End);
+	
 
 	////////////////////////////////////////////	画秒针的短轴	
-	m_s = (m_s + 30) % 60;
 	S   = (S + 30) % 60;
-	GetRgb(m_s, tRgb);
+	
+	if ( !m_bFirstClock )
+	{
+		//擦除上次的指针
+		cRgb = m_cLastSecS;
+		m_Point_End = m_point_lastSecS;
 
-	PenNew.DeleteObject();
-	PenNew.CreatePen(PS_DASHDOTDOT, 2, RGB(tRgb.x, tRgb.y, tRgb.z));
-	pPenOld = dc.SelectObject(&PenNew);
-	m_Point_End.x = 65 + LONG(6 * sin(m_s*PI / 30));
-	m_Point_End.y = 64 - LONG(6 * cos(m_s*PI / 30));
-	dc.MoveTo(m_Point_Start);
-	dc.LineTo(m_Point_End);
+		PenNew.DeleteObject();
+		PenNew.CreatePen(PS_DASHDOTDOT, 2, cRgb);
+		pPenOld = dc.SelectObject(&PenNew);
+		dc.MoveTo(m_Point_Start);
+		dc.LineTo(m_Point_End);
+	}
 
 
+	//画秒针短轴
 	PenNew.DeleteObject();
 	PenNew.CreatePen(PS_DASHDOTDOT, 2, RGB(255, 0, 0));
 	pPenOld = dc.SelectObject(&PenNew);
 	m_Point_End.x = 65 + LONG(6 * sin(S*PI / 30));
 	m_Point_End.y = 64 - LONG(6 * cos(S*PI / 30));
+
+	m_cLastSecS = dc.GetPixel(m_Point_End);
+	m_point_lastSecS = m_Point_End;
+
 	dc.MoveTo(m_Point_Start);
 	dc.LineTo(m_Point_End);
+	
 
 
 	///////////////////////画秒针的长轴
-	m_s = (m_s + 30) % 60;
 	S   = (S + 30) % 60;
-	GetRgb(m_s, tRgb);
+	
+	if ( !m_bFirstClock )
+	{
+		//擦除上次的指针
+		cRgb = m_cLastSecL;
+		m_Point_End = m_point_lastSecL;
 
-	//		BrushNew.DeleteObject();
-	//		BrushNew.CreateSolidBrush(RGB(C,C,C));
-	//		BrushOld=dc.SelectObject(&BrushNew);
-	PenNew.DeleteObject();
-	PenNew.CreatePen(PS_DASHDOTDOT, 2, RGB(tRgb.x, tRgb.y, tRgb.z));
-	pPenOld = dc.SelectObject(&PenNew);
-	m_Point_End.x = 65 + LONG(30 * sin(m_s*PI / 30));
-	m_Point_End.y = 64 - LONG(30 * cos(m_s*PI / 30));
-	dc.MoveTo(m_Point_Start);
-	dc.LineTo(m_Point_End);
+		//		BrushNew.DeleteObject();
+		//		BrushNew.CreateSolidBrush(RGB(C,C,C));
+		//		BrushOld=dc.SelectObject(&BrushNew);
+		PenNew.DeleteObject();
+		PenNew.CreatePen(PS_DASHDOTDOT, 2, cRgb);
+		pPenOld = dc.SelectObject(&PenNew);
+		dc.MoveTo(m_Point_Start);
+		dc.LineTo(m_Point_End);
+	}
 
 
+	//画秒针长轴
 	//		BrushNew.DeleteObject();
 	//		BrushNew.CreateSolidBrush(RGB(255,0,0));
 	//		BrushOld=dc.SelectObject(&BrushNew);
@@ -414,14 +439,18 @@ void  CAngoTimeDlg::ClockTime()
 	pPenOld = dc.SelectObject(&PenNew);
 	m_Point_End.x = 65 + LONG(30 * sin(S*PI / 30));
 	m_Point_End.y = 64 - LONG(30 * cos(S*PI / 30));
+
+	m_cLastSecL = dc.GetPixel(m_Point_End);
+	m_point_lastSecL = m_Point_End;
+
 	dc.MoveTo(m_Point_Start);
 	dc.LineTo(m_Point_End);
 
 
 
 	//////////////////////////////////////////////////////////////////
-	m_h = H; m_m = M; m_s = S;
-	//////////////////////////////////////////////////////////////////
+	m_bFirstClock = FALSE;
+	
 	dc.SetPixel(m_Point_Start, RGB(0, 0, 0));
 	dc.SetPixel(m_Point_Start.x + 1, m_Point_Start.y,	  RGB(0, 0, 0));
 	dc.SetPixel(m_Point_Start.x,	 m_Point_Start.y + 1, RGB(0, 0, 0));
@@ -430,6 +459,11 @@ void  CAngoTimeDlg::ClockTime()
 	dc.SetPixel(m_Point_Start.x,	 m_Point_Start.y - 1, RGB(0, 0, 0));
 	dc.SetPixel(m_Point_Start.x - 1, m_Point_Start.y - 1, RGB(0, 0, 0));
 }
+
+
+//-------------------------------------------------------------------------------------------------------------------------
+
+
 
 void CAngoTimeDlg::OnTimer(UINT_PTR nIDEvent)//控制时钟走动
 {
@@ -568,4 +602,285 @@ void CAngoTimeDlg::OnMenuAngo()
 // 	si.wShowWindow = SW_SHOW;
 // 	si.dwFlags = STARTF_USESHOWWINDOW;
 // 	BOOL fRet = CreateProcess(L"Ango.exe", NULL, NULL, FALSE, NULL, NULL, NULL, NULL, &si, &pi);
+}
+
+void CAngoTimeDlg::OnMenuExit()
+{
+	OnCancel();
+}
+
+//-------------------------------------------------------------------------------------------------------------------------
+void CAngoTimeDlg::InitThread()
+{
+	if ( g_pthAlarm || g_pthSayTime || g_pthWork )
+		return;
+
+	//创建报时线程
+	DWORD dwParam;
+	g_pthAlarm = AfxBeginThread(ThProcAlarm, &dwParam, THREAD_PRIORITY_HIGHEST, 0, CREATE_SUSPENDED, 0);
+	if (g_pthAlarm == NULL)
+		AngoMessageBox(L"创建线程错误!");
+
+
+	//创建闹钟线程
+	DWORD dwParam2;
+	g_pthSayTime = AfxBeginThread(ThProcSayTime, &dwParam2, THREAD_PRIORITY_HIGHEST, 0, CREATE_SUSPENDED, 0);
+	if (g_pthSayTime == NULL)
+		AngoMessageBox(L"创建线程错误!");
+
+
+	//创建处理定时任务线程
+	DWORD dwParam3;
+	g_pthWork = AfxBeginThread(ThProcWork, &dwParam3, THREAD_PRIORITY_HIGHEST, 0, CREATE_SUSPENDED, 0);
+	if (g_pthWork == NULL)
+		AngoMessageBox(L"创建线程错误!");
+
+	
+}
+
+UINT ThProcAlarm(LPVOID pParam)
+{
+	CString strPath;
+	while (g_bWork)
+	{
+		g_pthSayTime->SuspendThread();//挂起其它两个线程
+		g_pthWork->SuspendThread();
+		strPath = "c:\\win95\\media\\The Microsoft Sound.wav";
+		PlaySound(strPath, NULL, SND_FILENAME | SND_ASYNC);
+// 		PlaySound(MAKEINTRESOURCE(IDR_WAVESOUND), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+// 		Sleep(7000);
+// 		PlaySound(MAKEINTRESOURCE(IDR_WAVESOUND), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+// 		Sleep(8000);
+		g_pthSayTime->ResumeThread();//恢复其它线程
+		g_pthWork->ResumeThread();
+		g_pthAlarm->SuspendThread();
+	}
+	TRACE("exit ThProcAlarm\n");
+	return 0;
+}
+UINT ThProcSayTime(LPVOID pParam)
+{
+	while (g_bWork)
+	{
+		//挂起其它线程
+		g_pthWork->SuspendThread();
+		g_pthAlarm->SuspendThread();
+
+		CTime m_NowTime;
+		m_NowTime = CTime::GetCurrentTime();
+		int n_Hour = m_NowTime.GetHour();
+		//以下为报时代码,原理为通过对当前时间的判断,调用不同的声音文件(资源文件)进行组合播放,形成报时效果
+		PlaySound(MAKEINTRESOURCE(IDR_WAVE_TIMENOW), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+		Sleep(1550);
+		if (n_Hour < 8)
+			PlaySound(MAKEINTRESOURCE(IDR_WAVE_MORNING), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+		else if (n_Hour < 13)
+			PlaySound(MAKEINTRESOURCE(IDR_WAVE_AM), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+		else if (n_Hour < 19)
+			PlaySound(MAKEINTRESOURCE(IDR_WAVE_PM), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+		else
+			PlaySound(MAKEINTRESOURCE(IDR_WAVE_EM), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+		if (n_Hour > 12)
+		{
+			n_Hour = n_Hour - 12;
+		}
+		int n_Minute = m_NowTime.GetMinute();
+		int n_Second = m_NowTime.GetSecond();
+
+
+
+		Sleep(580);
+		switch (n_Hour)
+		{
+		case 0:
+			PlaySound(MAKEINTRESOURCE(IDR_WAVE_T00), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+			break;
+		case 1:
+			PlaySound(MAKEINTRESOURCE(IDR_WAVE_T01), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+			break;
+		case 2:
+			PlaySound(MAKEINTRESOURCE(IDR_WAVE_T2), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+			break;
+		case 3:
+			PlaySound(MAKEINTRESOURCE(IDR_WAVE_T03), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+			break;
+		case 4:
+			PlaySound(MAKEINTRESOURCE(IDR_WAVE_T04), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+			break;
+		case 5:
+			PlaySound(MAKEINTRESOURCE(IDR_WAVE_T05), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+			break;
+		case 6:
+			PlaySound(MAKEINTRESOURCE(IDR_WAVE_T06), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+			break;
+		case 7:
+			PlaySound(MAKEINTRESOURCE(IDR_WAVE_T07), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+			break;
+		case 8:
+			PlaySound(MAKEINTRESOURCE(IDR_WAVE_T08), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+			break;
+		case 9:
+			PlaySound(MAKEINTRESOURCE(IDR_WAVE_T09), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+			break;
+		case 10:
+			PlaySound(MAKEINTRESOURCE(IDR_WAVE_T10), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+			break;
+		case 11:
+			PlaySound(MAKEINTRESOURCE(IDR_WAVE_T11), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+			break;
+		case 12:
+			PlaySound(MAKEINTRESOURCE(IDR_WAVE_T12), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+			break;
+		}
+		if (n_Hour < 10)
+			Sleep(450);
+		else
+			Sleep(550);
+		PlaySound(MAKEINTRESOURCE(IDR_WAVE_POINT), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+		Sleep(500);
+
+		int n_Front_Min;
+		int n_Behind_Min;
+		int n_FTemp = n_Minute;
+
+		n_Front_Min = n_Minute / 10;
+		n_Behind_Min = n_Minute % 10;
+
+		if (n_Front_Min <= 0)
+		{
+			if (n_FTemp != 0)
+				PlaySound(MAKEINTRESOURCE(IDR_WAVE_T00), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+			Sleep(400);
+			switch (n_FTemp)
+			{
+			case 0:
+				break;
+			case 1:
+				PlaySound(MAKEINTRESOURCE(IDR_WAVE_T01), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+				break;
+			case 2:
+				PlaySound(MAKEINTRESOURCE(IDR_WAVE_T02), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+				break;
+			case 3:
+				PlaySound(MAKEINTRESOURCE(IDR_WAVE_T03), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+				break;
+			case 4:
+				PlaySound(MAKEINTRESOURCE(IDR_WAVE_T04), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+				break;
+			case 5:
+				PlaySound(MAKEINTRESOURCE(IDR_WAVE_T05), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+				break;
+			case 6:
+				PlaySound(MAKEINTRESOURCE(IDR_WAVE_T06), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+				break;
+			case 7:
+				PlaySound(MAKEINTRESOURCE(IDR_WAVE_T07), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+				break;
+			case 8:
+				PlaySound(MAKEINTRESOURCE(IDR_WAVE_T08), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+				break;
+			case 9:
+				PlaySound(MAKEINTRESOURCE(IDR_WAVE_T09), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+				break;
+			default:
+				break;
+			}
+			Sleep(500);
+			if (n_FTemp != 0)
+				PlaySound(MAKEINTRESOURCE(IDR_WAVE_MIN), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+
+		}
+		else
+		{
+			switch (n_Front_Min)
+			{
+			case 1:
+				PlaySound(MAKEINTRESOURCE(IDR_WAVE_T10), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+				break;
+			case 2:
+				PlaySound(MAKEINTRESOURCE(IDR_WAVE_T20), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+				break;
+			case 3:
+				PlaySound(MAKEINTRESOURCE(IDR_WAVE_T30), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+				break;
+			case 4:
+				PlaySound(MAKEINTRESOURCE(IDR_WAVE_T40), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+				break;
+			case 5:
+				PlaySound(MAKEINTRESOURCE(IDR_WAVE_T50), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+				break;
+			}
+			if (n_Front_Min == 1)
+				Sleep(470);
+			else if (n_Front_Min == 4)
+				Sleep(565);
+			else if (n_Behind_Min > 0)
+				Sleep(520);
+			switch (n_Behind_Min)
+			{
+			case 1:
+				PlaySound(MAKEINTRESOURCE(IDR_WAVE_T01), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+				break;
+			case 2:
+				PlaySound(MAKEINTRESOURCE(IDR_WAVE_T02), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+				break;
+			case 3:
+				PlaySound(MAKEINTRESOURCE(IDR_WAVE_T03), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+				break;
+			case 4:
+				PlaySound(MAKEINTRESOURCE(IDR_WAVE_T04), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+				break;
+			case 5:
+				PlaySound(MAKEINTRESOURCE(IDR_WAVE_T05), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+				break;
+			case 6:
+				PlaySound(MAKEINTRESOURCE(IDR_WAVE_T06), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+				break;
+			case 7:
+				PlaySound(MAKEINTRESOURCE(IDR_WAVE_T07), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+				break;
+			case 8:
+				PlaySound(MAKEINTRESOURCE(IDR_WAVE_T08), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+				break;
+			case 9:
+				PlaySound(MAKEINTRESOURCE(IDR_WAVE_T09), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+				break;
+			}
+
+			Sleep(560);
+			PlaySound(MAKEINTRESOURCE(IDR_WAVE_MIN), AfxGetApp()->m_hInstance, SND_RESOURCE | SND_ASYNC | SND_NODEFAULT);
+		}
+
+
+		//恢复其它线程，挂起本线程
+		g_pthWork->ResumeThread();
+		g_pthAlarm->ResumeThread();
+		g_pthSayTime->SuspendThread();
+
+	}
+
+	TRACE("exit ThProcSayTime\n");
+	return 0;
+}
+UINT ThProcWork(LPVOID pParam)
+{
+	while (g_bWork)
+	{
+		g_pthWork->SuspendThread();
+	}
+
+	TRACE("exit ThProcWork\n");
+	return 0;
+}
+
+
+
+
+void CAngoTimeDlg::OnSaytimeNow()
+{
+	if (g_pthSayTime)
+	{
+		g_pthSayTime->ResumeThread();
+	}
+
 }
